refactor: update implementation and doc with suggestions
