+++ Cargo.toml
@@ -14,8 +14,8 @@ members = [
   "examples/counter",
   "examples/dyn_create_destroy_apps",
   "examples/file_upload",
-  "examples/function_todomvc",
   "examples/function_memory_game",
+  "examples/function_todomvc",
   "examples/futures",
   "examples/game_of_life",
   "examples/inner_html",
