+++ examples/function_memory_game/index.html
@@ -4,8 +4,8 @@
     <meta charset="utf-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1" />
     <link rel="icon" href="public/favicon.ico" />
-    <title>Yew • Function MemoryGame</title>
-    <base data-trunk-public-url/>
+    <title>Yew • Function Memory Game</title>
+    <base data-trunk-public-url />
 
     <link data-trunk rel="sass" href="scss/index.scss" />
     <link data-trunk rel="copy-dir" href="public/" />
