+++ examples/function_memory_game/src/components.rs
@@ -1,3 +1,4 @@
+pub mod app;
 pub mod chessboard;
 pub mod chessboard_card;
 pub mod game_status_board;
