+++ examples/function_memory_game/src/components/score_board.rs
@@ -1,4 +1,4 @@
-use yew::{function_component, html, Properties};
+use yew::{function_component, html, Html, Properties};
 
 use crate::components::{
     score_board_best_score::BestScore, score_board_logo::Logo, score_board_progress::GameProgress,
@@ -10,13 +10,17 @@ pub struct Props {
     pub best_score: u32,
 }
 
-#[function_component(ScoreBoard)]
-pub fn score_board(props: &Props) -> Html {
+#[function_component]
+pub fn ScoreBoard(props: &Props) -> Html {
+    let Props {
+        best_score,
+        unresolved_card_pairs,
+    } = props.clone();
     html! {
         <div class="score-board">
             <Logo />
-            <GameProgress unresolved_card_pairs={props.unresolved_card_pairs} />
-            <BestScore best_score={props.best_score} />
+            <GameProgress {unresolved_card_pairs} />
+            <BestScore {best_score} />
         </div>
     }
 }
