+++ examples/function_memory_game/src/components/score_board_best_score.rs
@@ -1,12 +1,12 @@
-use yew::{function_component, html, Properties};
+use yew::{function_component, html, Html, Properties};
 
 #[derive(PartialEq, Properties, Clone)]
 pub struct Props {
     pub best_score: u32,
 }
 
-#[function_component(BestScore)]
-pub fn score_board_best_score(props: &Props) -> Html {
+#[function_component]
+pub fn BestScore(props: &Props) -> Html {
     html! {
         <div class="best-score">
             <span>{"Highest Record"}</span>
