+++ examples/function_memory_game/src/components/score_board_logo.rs
@@ -1,7 +1,7 @@
-use yew::{function_component, html};
+use yew::{function_component, html, Html};
 
-#[function_component(Logo)]
-pub fn score_board_logo() -> Html {
+#[function_component]
+pub fn Logo() -> Html {
     html! {
         <h1 class="logo">
             <a href="https://examples.yew.rs/function_memory_game" target="_blank">{"Memory"}</a>
